+++ src/main/java/com/ibm/restconf/config/RCDriverConstants.java
@@ -7,8 +7,6 @@ public abstract class RCDriverConstants {
     public static final String AUTHENTICATION_USERNAME = "username";
     public static final String AUTHENTICATION_PASSWORD = "password";
     public static final String AUTHENTICATION_URL = "authenticationUrl";
-    public static final String AUTHENTICATION_USERNAME_TOKEN_NAME = "usernameTokenName";
-    public static final String AUTHENTICATION_PASSWORD_TOKEN_NAME = "passwordTokenName";
     public static final String AUTHENTICATION_CLIENT_ID = "client_id";
     public static final String AUTHENTICATION_CLIENT_SECRET = "client_secret";
     public static final String AUTHENTICATION_ACCESS_TOKEN_URI = "accessTokenUri";