+++ src/main/java/com/ibm/restconf/service/impl/CiscoCncAuthService.java
@@ -3,7 +3,6 @@ package com.ibm.restconf.service.impl;
 import com.ibm.restconf.driver.CiscoCncServiceDriver;
 import com.ibm.restconf.model.ExecutionRequest;
 import com.ibm.restconf.service.AuthService;
-import com.ibm.restconf.service.LifecycleManagementService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,7 +13,7 @@ public class CiscoCncAuthService implements AuthService {
 
     private final static Logger logger = LoggerFactory.getLogger(CiscoCncAuthService.class);
 
-    private CiscoCncServiceDriver ciscoCncAuthServiceDriver;
+    private final CiscoCncServiceDriver ciscoCncAuthServiceDriver;
 
     @Autowired
     public CiscoCncAuthService(CiscoCncServiceDriver ciscoCncAuthServiceDriver){
@@ -22,17 +21,17 @@ public class CiscoCncAuthService implements AuthService {
     }
 
     @Override
-    public String authenticate(ExecutionRequest executionRequest) {
+    public String authenticate(ExecutionRequest executionRequest, String driverRequestId) {
         logger.info("Calling Cisco CNC API to get JWT Token.");
-        return getToken(executionRequest, getTicket(executionRequest));
+        return getToken(executionRequest, getTicket(executionRequest, driverRequestId), driverRequestId);
     }
 
-    public String getTicket(ExecutionRequest executionRequest){
+    public String getTicket(ExecutionRequest executionRequest, String driverRequestId){
         logger.info("Calling Cisco CNC API to get Ticket.");
-        return this.ciscoCncAuthServiceDriver.getTicket(executionRequest.getDeploymentLocation());
+        return this.ciscoCncAuthServiceDriver.getTicket(executionRequest.getDeploymentLocation(), driverRequestId);
     }
-    private String getToken(ExecutionRequest executionRequest, String ticket){
-        return this.ciscoCncAuthServiceDriver.getToken(executionRequest.getDeploymentLocation(), ticket);
+    private String getToken(ExecutionRequest executionRequest, String ticket, String driverRequestId){
+        return this.ciscoCncAuthServiceDriver.getToken(executionRequest.getDeploymentLocation(), ticket, driverRequestId);
     }
 
 }