+++ src/main/java/com/ibm/restconf/service/impl/JinJavaMessageConversionServiceImpl.java
@@ -15,8 +15,6 @@ import org.springframework.util.StringUtils;
 import java.io.*;
 import java.nio.charset.Charset;
 import java.util.*;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 @Service("JinJavaMessageConversionServiceImpl")
 public class JinJavaMessageConversionServiceImpl implements MessageConversionService {
@@ -29,7 +27,7 @@ public class JinJavaMessageConversionServiceImpl implements MessageConversionSer
         Map<String, Object> resoureceProperties = executionRequest.getProperties();
         String contentType = (String)resoureceProperties.get("ContentType");
         String fullTemplateName = messageType + ".xml";
-        if(!StringUtils.isEmpty(contentType) && contentType.equalsIgnoreCase("json")){
+        if(StringUtils.hasLength(contentType) && contentType.equalsIgnoreCase("json")){
             fullTemplateName = messageType+ ".json";
         }
         logger.debug("fullTemplateName  {} \n", fullTemplateName);
@@ -37,7 +35,7 @@ public class JinJavaMessageConversionServiceImpl implements MessageConversionSer
         logger.debug("template file from ExecutionRequest {} \n", template);
         final Jinjava jinjava = new Jinjava();
         try {
-            String returnVal = jinjava.render(template, createJinJavaContext(executionRequest.getProperties(), template));
+            String returnVal = jinjava.render(template, createJinJavaContext(executionRequest.getProperties()));
             logger.info("Message conversion script successfully run, returnVal is\n{}", returnVal);
             return returnVal;
         } catch (IOException e) {
@@ -67,9 +65,9 @@ public class JinJavaMessageConversionServiceImpl implements MessageConversionSer
     }
 
 
-    private Map<String,Object> createJinJavaContext(Map<String, Object> resourceProperties, String templateContents) throws  IOException {
+    private Map<String,Object> createJinJavaContext(Map<String, Object> resourceProperties) throws  IOException {
         Map<String, Object> context = new HashMap<>();
-        List<String> list = findPropertyListFromTemplate(templateContents);
+        Set<String> list = resourceProperties.keySet();
         logger.debug("list of properties in template {}", list);
         list.forEach(property -> {
             Object value = resourceProperties.get(property);
@@ -80,22 +78,5 @@ public class JinJavaMessageConversionServiceImpl implements MessageConversionSer
             context.put(property, value);
         });
         return context;
-    }
-
-    private List<String> findPropertyListFromTemplate(String templateContents) {
-        List<String> list = new ArrayList<>();
-        Pattern pattern = Pattern.compile("\\{\\{(.*?)}}");
-        Matcher match = pattern.matcher(templateContents);
-        while(match.find()) {
-            int start = match.start(0);
-            int end = match.end(0);
-            String extractedProperty = templateContents.substring(start + 2, end - 2);
-            logger.debug("Extracted property {}", extractedProperty);
-            list.add(extractedProperty);
-        }
-        if(list.isEmpty()){
-            logger.debug("No properties extracted from the template {}", templateContents);
-        }
-        return list;
     }
 }